code/ai-agent/requirements.txt
google-cloud-aiplatform[reasoningengine,adk]>=1.55.0
langchain-google-vertexai
langchain-community
langchain-classic